[KYUUBI #1080] Move stringifyException to Utils
